% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{codes_tab}
\alias{codes_tab}
\title{Make a table of ICD/CPT/etc codes from lists}
\usage{
codes_tab(..., sep = ", ")
}
\arguments{
\item{...}{lists of codes used to identify different diagnoses, procedures, etc.
Different coding systems (ICD9, ICD10, CPT, ...) would usually get separate lists.}

\item{sep}{codes are pasted together into a single cell separated by this string (default = ", ")}
}
\value{
data.frame with one row per diagnosis/procedure (variable column) and a column for each list provided (coding system).
}
\description{
Make a table of ICD/CPT/etc codes from lists
}
