% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{extract_time}
\alias{extract_time}
\title{extract times from pasted cells}
\usage{
extract_time(
  x,
  lower,
  upper,
  extract = c("between", "min", "max"),
  sep = "\\\\|"
)
}
\arguments{
\item{x}{column of times pasted together ("-1 | 0 | 1")}

\item{lower}{min value of interest}

\item{upper}{max value of interest}

\item{extract}{between = any values in x between lower and upper, min/max = the lowest or largest number between lower and upper}

\item{sep}{character separating times in x}
}
\value{
vector of T/F or min/max times (if any)
}
\description{
extract times from pasted cells
}
