% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{factorn}
\alias{factorn}
\title{Convert to factor with levels ordered by frequency}
\usage{
factorn(x, decreasing = TRUE)
}
\arguments{
\item{x}{vector of values (preferably character)}

\item{decreasing}{TRUE (default) = factor levels in decreasing order of frequency, FALSE = increasing}
}
\value{
factor with levels in order of frequency
}
\description{
Convert to factor with levels ordered by frequency
}
