% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{fill}
\alias{fill}
\title{fill up or down}
\usage{
fill(x, direction = c("down", "up"))
}
\arguments{
\item{x}{vector with missing entries}

\item{direction}{fill up or down}
}
\value{
vector with elements filled in
}
\description{
fill up or down
}
