% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{get_fda}
\alias{get_fda}
\title{Query openFDA for vector of drug names}
\usage{
get_fda(druglist, route = NULL)
}
\arguments{
\item{druglist}{list of drug names (this is a wrapper for query_fda). See example}

\item{route}{optional route of administration (same for all in druglist)}
}
\value{
list of results
}
\description{
Query openFDA for vector of drug names
}
\examples{
get_fda(list("oxybutynin" = "oxybutynin"), route = NULL) # |> get_ndc()
# more than one generic name could be provided in the list
# (e.g. list("global_name" = c("name1", "name2")))
}
