% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{get_ndc}
\alias{get_ndc}
\title{Extract NDCs from results of get_fda}
\usage{
get_ndc(fda)
}
\arguments{
\item{fda}{output from get_fda}
}
\value{
list of NDCs
}
\description{
Extract NDCs from results of get_fda
}
\examples{
get_fda(list("oxybutynin" = "oxybutynin"), route = NULL) |>
get_ndc()
}
