% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{get_rxcui}
\alias{get_rxcui}
\title{Extract RxCUI from results of get_fda}
\usage{
get_rxcui(fda)
}
\arguments{
\item{fda}{output from get_fda}
}
\value{
list of RxCUI
}
\description{
Extract RxCUI from results of get_fda
}
\examples{
get_fda(list("oxybutynin" = "oxybutynin"), route = NULL) |>
get_rxcui()
}
