% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{get_specs}
\alias{get_specs}
\title{Get NIS file specifications}
\usage{
get_specs(year, file = c("Core", "Hospital", "Severity"))
}
\arguments{
\item{year}{of interest}

\item{file}{"Core", "Hospital", or "Severity"}
}
\value{
data.frame with 4 columns ("type", "varname", "start", "end") for use in reading ASC files
}
\description{
Uses stata load files from hcup-us.ahrq.gov to get specs for loading NIS files
}
