% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdfun.R
\name{icd10_convert}
\alias{icd10_convert}
\title{Convert ICD10 codes from short to decimal or back}
\usage{
icd10_convert(x, direction = c("short", "decimal"))
}
\arguments{
\item{x}{vector of ICD 10 codes}

\item{direction}{short or decimal}
}
\value{
vector of converted codes (not checked for existence!)
}
\description{
Convert ICD10 codes from short to decimal or back
}
