% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdfun.R
\name{icd9_convert}
\alias{icd9_convert}
\title{Convert ICD9 codes from short to decimal or back}
\usage{
icd9_convert(x, direction = c("short", "decimal"))
}
\arguments{
\item{x}{vector of ICD 9 codes}

\item{direction}{short or decimal}
}
\value{
vector of converted codes (not checked for existence!)
}
\description{
Convert ICD9 codes from short to decimal or back
}
