% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{load_nis}
\alias{load_nis}
\title{Load NIS files into R}
\usage{
load_nis(year, file = c("Core", "Hospital", "Severity"), nis_path)
}
\arguments{
\item{year}{of interest}

\item{file}{"Core", "Hospital", or "Severity"}

\item{nis_path}{path to data files. If not given, assume they are in a subfolder titled paste0("NIS_", year)}
}
\value{
data.frame containing NIS data
}
\description{
Load NIS files into R
}
