% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{mkorder}
\alias{mkorder}
\title{Paste order of events together}
\usage{
mkorder(data, cols, labels, asfactor = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{cols}{column names in data (numeric or NA). Columns must be numeric}

\item{labels}{optional names to substitute for cols}

\item{asfactor}{return factor (with levels ordered by count) or not}
}
\value{
character vector of orders where x->y means y followed x and x=y means x
and y co-occured.
}
\description{
Paste order of events together
}
\examples{
d <- data.frame(x1 = c(-10, 0, NA), x2 = c(NA, -10, 30), x3 = c(10, 20, 30))

srmi::mkorder(d, cols = colnames(d), labels = c("A", "B", "C"), asfactor = FALSE)
}
