% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{ndc10to11}
\alias{ndc10to11}
\title{Convert 10 digit NDCs to 11 digit}
\usage{
ndc10to11(x)
}
\arguments{
\item{x}{a 10 digit NDC code (character) with dashes "-". Should be 4-4-2, 5-3-2, or 5-4-1.}
}
\value{
11 digit NDC
}
\description{
Convert 10 digit NDCs to 11 digit
}
