% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmi.R
\name{query_fda}
\alias{query_fda}
\title{Query openFDA by drug name}
\usage{
query_fda(dname, route)
}
\arguments{
\item{dname}{drug name (generic)}

\item{route}{optional route of administration}
}
\value{
data.frame of results for processing by other functions
}
\description{
Query openFDA by drug name
}
