% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdfun.R
\name{children}
\alias{children}
\title{Find children (sub-codes) of ICD9/10 codes using `icd.data`}
\usage{
children(x, format = c("short", "decimal"), unlist = FALSE)
}
\arguments{
\item{x}{vector of ICD9 or 10 diagnosis codes (will guess based on starting character)}

\item{format}{of the returned codes (short or decimal)}

\item{unlist}{return a list of codes, with a slot for each code in x, or not}
}
\value{
a list or vector of subcodes
}
\description{
Find children (sub-codes) of ICD9/10 codes using `icd.data`
}
